Adding working signup and login form views.
