--- ios/broccoli/SecondViewController.swift
@@ -1,17 +1,0 @@
-//
-//  SecondViewController.swift
-//  broccoli
-//
-//  Created by Nuno on 17/04/2017.
-//  Copyright © 2017 nunoalexandre. All rights reserved.
-//
-
-import UIKit
-
-class SecondViewController: UIViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
-
